fastapi_backend/requirements.txt
fastapi>=0.68.0,<1.0.0
pydantic>=1.8.0,<2.0.0
langchain>=0.0.247
Pillow>=9.0.0
python-dotenv>=0.19.0
pymysql>=1.0.2
torch>=2.0.0
transformers>=4.30.0
rembg==2.0.50
uvicorn>=0.15.0
langchain_openai>=0.0.2
python-multipart>=0.0.5
mangum>=0.17.0
numpy>=1.21.0